# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UrbEx Explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface RegisterBody {
  username: string;
  password: string;
  email: string;
  bio?: string;
  location?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  approved: "approved",
  banned: "banned",
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  bio?: string;
  location?: string;
  joinedAt: string;
  postCount?: number;
  locationCount?: number;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type UpdateUserStatusBodyStatus =
  (typeof UpdateUserStatusBodyStatus)[keyof typeof UpdateUserStatusBodyStatus];

export const UpdateUserStatusBodyStatus = {
  pending: "pending",
  approved: "approved",
  banned: "banned",
} as const;

export type UpdateUserStatusBodyRole =
  (typeof UpdateUserStatusBodyRole)[keyof typeof UpdateUserStatusBodyRole];

export const UpdateUserStatusBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UpdateUserStatusBody {
  status?: UpdateUserStatusBodyStatus;
  role?: UpdateUserStatusBodyRole;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  color: string;
  description?: string;
  locationCount?: number;
}

export type LocationRiskLevel =
  (typeof LocationRiskLevel)[keyof typeof LocationRiskLevel];

export const LocationRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  extreme: "extreme",
} as const;

export type LocationStatus =
  (typeof LocationStatus)[keyof typeof LocationStatus];

export const LocationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Location {
  id: number;
  title: string;
  description?: string;
  latitude: number;
  longitude: number;
  categoryId: number;
  categoryName?: string;
  categoryIcon?: string;
  categoryColor?: string;
  state?: string;
  country?: string;
  addedById?: number;
  addedByUsername?: string;
  riskLevel: LocationRiskLevel;
  status: LocationStatus;
  imageUrl?: string;
  createdAt: string;
  viewCount?: number;
}

export interface Thread {
  id: number;
  title: string;
  content: string;
  authorId: number;
  authorUsername: string;
  forumCategoryId?: number;
  forumCategoryName?: string;
  postCount?: number;
  viewCount?: number;
  isPinned?: boolean;
  isLocked?: boolean;
  lastPostAt?: string;
  lastPostUsername?: string;
  createdAt: string;
}

export type LocationDetail = Location & {
  images?: string[];
  accessNotes?: string;
  hazards?: string;
  bestTimeToVisit?: string;
  relatedThreads?: Thread[];
};

export type CreateLocationBodyRiskLevel =
  (typeof CreateLocationBodyRiskLevel)[keyof typeof CreateLocationBodyRiskLevel];

export const CreateLocationBodyRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  extreme: "extreme",
} as const;

export interface CreateLocationBody {
  title: string;
  description: string;
  latitude: number;
  longitude: number;
  categoryId: number;
  state?: string;
  country?: string;
  riskLevel: CreateLocationBodyRiskLevel;
  imageUrl?: string;
  accessNotes?: string;
  hazards?: string;
  bestTimeToVisit?: string;
}

export type UpdateLocationBodyRiskLevel =
  (typeof UpdateLocationBodyRiskLevel)[keyof typeof UpdateLocationBodyRiskLevel];

export const UpdateLocationBodyRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  extreme: "extreme",
} as const;

export type UpdateLocationBodyStatus =
  (typeof UpdateLocationBodyStatus)[keyof typeof UpdateLocationBodyStatus];

export const UpdateLocationBodyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateLocationBody {
  title?: string;
  description?: string;
  categoryId?: number;
  riskLevel?: UpdateLocationBodyRiskLevel;
  status?: UpdateLocationBodyStatus;
  accessNotes?: string;
  hazards?: string;
  bestTimeToVisit?: string;
  imageUrl?: string;
}

export type LocationStatsByCategoryItem = {
  categoryId: number;
  categoryName: string;
  categoryIcon?: string;
  categoryColor?: string;
  count: number;
};

export type LocationStatsByCountryItem = {
  country: string;
  count: number;
};

export interface LocationStats {
  total: number;
  byCategory: LocationStatsByCategoryItem[];
  recentlyAdded: number;
  byCountry: LocationStatsByCountryItem[];
}

export interface Post {
  id: number;
  content: string;
  authorId: number;
  authorUsername: string;
  authorRole?: string;
  authorJoinedAt?: string;
  authorPostCount?: number;
  threadId: number;
  createdAt: string;
  updatedAt?: string;
}

export type ThreadDetail = Thread & {
  posts?: Post[];
};

export interface CreateThreadBody {
  title: string;
  content: string;
  forumCategoryId: number;
}

export interface CreatePostBody {
  content: string;
}

export interface ForumCategory {
  id: number;
  name: string;
  description?: string;
  slug: string;
  icon?: string;
  threadCount?: number;
  postCount?: number;
  lastThreadTitle?: string;
  lastThreadAt?: string;
}

export interface ForumStats {
  totalThreads: number;
  totalPosts: number;
  totalMembers: number;
  newestMember?: string;
  onlineCount?: number;
}

export type ListUsersParams = {
  status?: ListUsersStatus;
};

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  pending: "pending",
  approved: "approved",
  banned: "banned",
} as const;

export type ListLocationsParams = {
  categoryId?: number;
  search?: string;
  state?: string;
};

export type ListThreadsParams = {
  categoryId?: number;
  search?: string;
  page?: number;
};