# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UrbEx Explorer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration submitted, pending admin approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /forum/threads:
    get:
      operationId: listThreads
      tags: [forum]
      summary: List forum threads
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Forum threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"
    post:
      operationId: createThread
      tags: [forum]
      summary: Create a new forum thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadBody"
      responses:
        "201":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
  /forum/threads/{id}:
    get:
      operationId: getThread
      tags: [forum]
      summary: Get a thread with all its posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thread with posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
    delete:
      operationId: deleteThread
      tags: [forum]
      summary: Delete a thread (admin or author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /forum/threads/{id}/posts:
    post:
      operationId: createPost
      tags: [forum]
      summary: Reply to a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /forum/posts/{id}:
    delete:
      operationId: deletePost
      tags: [forum]
      summary: Delete a post (admin or author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        bio:
          type: string
        location:
          type: string
      required: [username, password, email]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        status:
          type: string
          enum: [pending, approved, banned]
        bio:
          type: string
        location:
          type: string
        joinedAt:
          type: string
        postCount:
          type: integer
        locationCount:
          type: integer
      required: [id, username, email, role, status, joinedAt]
...
    Thread:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        forumCategoryId:
          type: integer
        forumCategoryName:
          type: string
        postCount:
          type: integer
        viewCount:
          type: integer
        isPinned:
          type: boolean
        isLocked:
          type: boolean
        lastPostAt:
          type: string
        lastPostUsername:
          type: string
        createdAt:
          type: string
      required: [id, title, content, authorId, authorUsername, createdAt]
    ThreadDetail:
      allOf:
        - $ref: "#/components/schemas/Thread"
        - type: object
          properties:
            posts:
              type: array
              items:
                $ref: "#/components/schemas/Post"
    CreateThreadBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        forumCategoryId:
          type: integer
      required: [title, content, forumCategoryId]
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorRole:
          type: string
        authorJoinedAt:
          type: string
        authorPostCount:
          type: integer
        threadId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, content, authorId, authorUsername, threadId, createdAt]
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...