# lib/api-zod/src/generated/types/createThreadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UrbEx Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateThreadBody {
  title: string;
  content: string;
  forumCategoryId: number;
}